Add a seperate config.py module
