+++ pyspectral/__init__.py
@@ -21,69 +21,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Main module"""
+"""Pyspectral package init"""
 
 from pyspectral.version import __version__
-import logging
-import os
-import yaml
-#from six.moves import configparser
-from collections import Mapping
-import pkg_resources
-
-LOG = logging.getLogger(__name__)
-
-BUILTIN_CONFIG_FILE = pkg_resources.resource_filename(__name__,
-                                                      os.path.join('etc', 'pyspectral.yaml'))
-
-CONFIG_FILE = os.environ.get('PSP_CONFIG_FILE')
-
-if CONFIG_FILE is not None and (not os.path.exists(CONFIG_FILE) or
-                                not os.path.isfile(CONFIG_FILE)):
-    raise IOError(str(CONFIG_FILE) + " pointed to by the environment " +
-                  "variable PSP_CONFIG_FILE is not a file or does not exist!")
-
-
-# def get_config():
-#     """Get config from file"""
-
-#     conf = configparser.ConfigParser()
-#     conf.read(BUILTIN_CONFIG_FILE)
-#     if CONFIG_FILE is not None:
-#         try:
-#             conf.read(CONFIG_FILE)
-#         except configparser.NoSectionError:
-#             LOG.info('Failed reading configuration file: %s',
-#                      str(CONFIG_FILE))
-
-#     return conf
-
-def recursive_dict_update(d, u):
-    """Recursive dictionary update using
-
-    Copied from:
-
-        http://stackoverflow.com/questions/3232943/update-value-of-a-nested-dictionary-of-varying-depth
-
-    """
-    for k, v in u.items():
-        if isinstance(v, Mapping):
-            r = recursive_dict_update(d.get(k, {}), v)
-            d[k] = r
-        else:
-            d[k] = u[k]
-    return d
-
-
-def get_config():
-    """Get the configuration from file"""
-    if CONFIG_FILE is not None:
-        configfile = CONFIG_FILE
-    else:
-        configfile = BUILTIN_CONFIG_FILE
-
-    config = {}
-    with open(configfile, 'r') as fp_:
-        config = recursive_dict_update(config, yaml.load(fp_))
-
-    return config
+#from pyspectral.config import get_config
